feat(bonuses): v0.2.5 conditional/stacking bonuses + atomicity + debt
Conditional bonus framework:
- ConditionalBonus model + ConditionType/BonusType enums (JSON-only contract)
- ConditionalBonusEvaluator static evaluator shared by gear + future legions
- 5 bonus types: FlatAttack, FlatDefense, ProcChanceFlat (clamped ≤1.0),
  ProcAmountFlat, FlatDamagePercent (applied after crit)
- 3 condition types: OwnedUnitCount, OwnedTypeCount, EquippedSlot
- GearDefinition.ConditionalBonuses; ItemDefinition.Tags
- EffectiveCombatData gains FlatDamagePercent; EquipmentService evaluates
  all equipped gear bonuses against player inventory (per-hit, indexed)

Debt fold-ins:
- Reward atomicity: stamina spend moved inside advisory-lock transaction
  (PlayerResourceRepository detects ambient tx); atomic rollback on race/failure,
  no refund path needed
- ProcBonus: double -> long in RaidHitResponse
- Function Reference fully refreshed to v0.2.5

Tests: 232 unit + 7 integration = 239 passing, 0 warnings

+++ src/ROTA.Shared/DTOs/RaidDTOs.cs
@@ -48,7 +48,7 @@ public class RaidHitResponse
     public bool IsCrit { get; set; }
     public double CritMultiplier { get; set; }
     public bool   ProcFired  { get; set; }
-    public double ProcBonus  { get; set; } // raw bonus damage from proc (0 if no proc)
+    public long   ProcBonus  { get; set; } // raw bonus damage from proc (0 if no proc)
 }
 
 public class RaidRewards
